moving hwm extended buffers to ds6 folder
